+++ AlgoritmDM/Com/Scenariy/NakopDM/FConfig.cs
@@ -97,7 +97,7 @@ namespace AlgoritmDM.Com.Scenariy.NakopDM
         /// </summary>
         /// <param name="DiscReasonId">Тип чека наличие которого нужно проверить.</param>
         /// <returns>Если существует в основном класса то возвращаем true</returns>
-        private bool HashDiscReasonId(int DiscReasonId)
+        private bool HashDiscReasonId(Int64 DiscReasonId)
         {
             bool rez = false;
             foreach (int item in this.MainNDM.NotDiscReasonId)
